requirements.txt
Flask==1.1.1
google-api-python-client
google-cloud-ndb
google-cloud-secret-manager
google-cloud-storage
google-cloud-translate

git+https://github.com/typeworld/typeworld.git
git+https://github.com/yanone/hotmetal.git

gunicorn
keyring
lxml
Markdown
markdown2
Pygments
pyjwt
pyzmq
semver

# https://stripe.com/docs/billing/subscriptions/metered
stripe>=2.60.0,<3.0
